Water level bar

+++ Assets/Scripts/Gameplay/DangerAreaSettings.cs
@@ -7,8 +7,8 @@ public class DangerAreaSettings : ScriptableObject
 {
     public int PiecesToStart { get => piecesToStart; }
     [SerializeField] int piecesToStart;
-    public float PiecesToMove { get => piecesToMove; }
-    [SerializeField] float piecesToMove;
+    public int PiecesToMove { get => piecesToMove; }
+    [SerializeField] int piecesToMove;
     public float TimeToMoveUp { get => timeToMoveUp; }
     [SerializeField] float timeToMoveUp;
     public float MaxFloor { get => maxFloor; }
